kernel: timeout: report the no-deadline case from the idle path
The idle path asks this for the time until the next wakeup, and gets a
tick count that never says "there is nothing to wake up for": an empty
timeout list is reported as the capped announce budget, exactly like a
deadline further out than can be programmed in one step. The power
management code cannot then tell the two apart, so it arms the timer in
both cases and a system with sloppy idle enabled keeps waking up for
nothing.

Make the same decision reprogram_next() makes. With
CONFIG_SYSTEM_CLOCK_SLOPPY_IDLE allowing the uptime to drift, an empty
list is reported as K_TICKS_FOREVER; anything else is a wait, whether a
real deadline or the synthetic one that keeps the announce range
covered. Without sloppy idle the empty case stays a wait, so the timer
remains armed and timekeeping is unaffected.

The return type becomes unsigned to match the tick type used throughout
the timer interface. The conversion at the only caller, in the idle
path, is value preserving in both directions: every wait is capped at
SYS_CLOCK_MAX_WAIT, which is INT32_MAX, and K_TICKS_FOREVER is the same
value read either way. Nothing outside the kernel sees the change; the
power management API keeps its signed tick counts.

+++ kernel/idle.c
@@ -54,7 +54,13 @@ void idle(void *unused1, void *unused2, void *unused3)
 		bool system_suspended = false;
 
 		if (!k_is_pre_kernel()) {
-			_kernel.idle = z_get_next_timeout_expiry();
+			/* _kernel.idle is int32_t and the helper is uint32_t.
+			 * Every value it returns is either at most
+			 * SYS_CLOCK_MAX_WAIT (== INT32_MAX) or the
+			 * K_TICKS_FOREVER bit pattern, which converts back
+			 * to -1, so nothing is lost here.
+			 */
+			_kernel.idle = (int32_t)z_get_next_timeout_expiry();
 
 			/*
 			 * Call the suspend hook function of the soc interface
